Added cell highlighting indicating brush position
Works with symmetry and different types of brushes.  In doing so, I
restructured the putpixel function, which now calls an internal function called
"_setPixel", which actually sets the values passed in, but checkes for special
"colours" like "highlight" and "cursor".  In those cases it adds the
appropriate class to the cell instead of actually changing it's colour.

This paves the way for drawing tools like "circle" and "line", which need a
highlighted indicator of where they will be placing pixels.